refactor: página de compra
