+++ src/main/webapp/paginas/compra.html
@@ -1,0 +1,68 @@
+<title>Compra</title>
+
+<nav id="main-nav"></nav>
+
+<div class="container py-5">
+    <h1 class="mb-4 text-center">Finalizar Compra</h1>
+
+    <!-- Seção de Endereço -->
+    <section id="secao-selecao-endereco" class="mb-4">
+        <label for="endereco" class="form-label">Selecione o Endereço:</label>
+        <select id="endereco" class="form-select"></select>
+
+        <a href="/conta">
+            <button class="btn btn-primary btn-sm mt-2">+ Adicionar novo endereço</button>
+        </a>
+    </section>
+
+    <!-- Seção de Cupons -->
+    <section id="secao-cupons" class="mb-4">
+        <p class="fw-bold">Adicionar Cupons</p>
+
+        <div id="cupons"></div>
+
+        <button id="btn-adicionar-cupom" class="btn btn-primary btn-sm mt-2">
+            + Adicionar cupom à compra
+        </button>
+        
+    </section>
+
+    <section id="secao-cartoes" class="mb-4">
+        <p class="fw-bold">Cartões de Crédito</p>
+
+        <div id="cartoes"></div>
+
+        <button id="btn-adicionar-cartao" class="btn btn-secondary btn-sm mt-2">
+            + Adicionar outro cartão
+        </button>
+
+        <a href="/conta">
+            <button id="btn-cadastrar-novo-cartao" class="btn btn-primary btn-sm mt-2">
+                + Cadastrar novo cartão
+            </button>
+        </a>
+    </section>
+
+    <!-- Resumo -->
+    <section id="resumo" class="mb-4">
+        <p class="fw-bold">Resumo</p>
+        <table id="tabela-produtos" class="table table-bordered table-hover">
+            <thead class="table-dark">
+                <th>Nome</th>
+                <th>Quantidade</th>
+                <th>Preço</th>
+                <th>Total</th>
+            </thead>
+            <tbody>
+
+            </tbody>
+        </table>
+    </section>
+
+    <!-- Botão Finalizar Compra -->
+    <div class="text-center">
+        <button id="btn-enviar-pedido" class="btn btn-primary btn-lg">Finalizar Compra</button>
+    </div>
+</div>
+
+<footer id="main-footer"></footer>
